refactor account switcher & management previews
